Move `Affine3DUtils` to `pyramid` package and update references
