+++ src/test/java/sc/fiji/ome/zarr/util/Affine3DUtilsTest.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -34,6 +34,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import net.imglib2.realtransform.AffineTransform3D;
 
 import org.junit.jupiter.api.Test;
+
+import sc.fiji.ome.zarr.pyramid.Affine3DUtils;
 
 /**
  * Unit tests for the {@link Affine3DUtils} class.
